pyproject.toml
[project]
name = "llm-gamebook"
version = "0.0.0"
description = "An interactive storytelling framework using LLMs to generate dynamic narratives"
license = "AGPL-3.0-or-later"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "casefy>=1.1.0",
  "colorama>=0.4.6",
  "jinja2>=3.1.6",
  "pydantic>=2.11.5",
  "pydantic-ai-slim[openai]>=0.2.9",
  "pyyaml>=6.0.2",
]

[project.urls]
source = "https://github.com/buzz/llm-gamebook"

[dependency-groups]
dev = [
  "mypy>=1.15.0",
  "pytest>=8.3.5",
  "ruff>=0.11.5",
  "types-colorama>=0.4.15.20240311",
  "types-pyyaml>=6.0.12.20250516",
]

[tool.ruff]
line-length = 100
output-format = "concise"
preview = true
exclude = [".venv/**"]

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules/
select = [
  "F",  # pyflakes
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "PL", # pylint

  # flake8
  "A",     #   -builtins
  "ASYNC", #   -async
  "B",     #   -bugbear
  "BLE",   #   -blind-except
  "C4",    #   -comprehensions
  "COM",   #   -commas
  "DTZ",   #   -datetimez
  "EM",    #   -errmsg
  "EXE",   #   -executable
  "FA",    #   -future-annotations
  "FBT",   #   -boolean-trap
  "G",     #   -logging-format
  "ICN",   #   -import-conventions
  "INP",   #   -no-pep420
  "INT",   #   -gettext
  "ISC",   #   -implicit-str-concat
  "LOG",   #   -logging
  "PIE",   #   -pie
  "PT",    #   -pytest-style
  "PYI",   #   -pyi
  "Q",     #   --quotes
  "RET",   #   -return
  "RSE",   #   -raise
  "SIM",   #   -simplify
  "SLOT",  #   -slots
  "T10",   #   -debugger
  "TCH",   #   -type-checking
  "TD004", #   -todos
  "TD005",
  "TD006",
  "TD007",
  "TID",   #   -tidy-imports
  "YTT",   #   -2020

  "FLY",  # flynt
  "FURB", # refurb
  "I",    # isort
  "N",    # pep8-naming
  "PERF", # Perflint
  "PGH",  # pygrep-hooks
  "RUF",  # Ruff-specific rules
  "TRY",  # tryceratops
  "UP",   # pyupgrade
]
ignore = ["COM812", "PLR1702", "PLR2004", "PLR6301", "RUF029"]

[tool.mypy]
disallow_untyped_defs = true
strict_optional = true
